Introduce a java 8 source set for Ratpack integration
