+++ dd-java-agent/instrumentation/ratpack-1.4/ratpack-1.4.gradle
@@ -9,37 +9,42 @@ versionScan {
   ]
 }
 
-if (JavaVersion.current().java8Compatible) {
-  apply from: "${rootDir}/gradle/java.gradle"
+apply from: "${rootDir}/gradle/java.gradle"
 
-  // Ratpack only supports Java 1.8+
+/*
+Here we introduce a sourceSet for the java 8 code which needs to be compiled with a source and target of 1.8
+The instrumentation classes must be compiled with java 7 and do nothing when ratpack is not on the classpath. The
+java 8 classes are used lazily so there is no direct linking between the 1.7 and 1.8 bytecode.
+*/
+sourceSets {
+  main_java8 {
+    java.srcDirs "${project.projectDir}/src/main/java8"
+  }
+}
+
+compileMain_java8Java {
   sourceCompatibility = 1.8
   targetCompatibility = 1.8
-  dependencies {
-    compileOnly group: 'io.ratpack', name: 'ratpack-core', version: '1.4.0'
-
-    compile project(':dd-trace-ot')
-    compile project(':dd-java-agent:agent-tooling')
-
-    compile deps.bytebuddy
-    compile deps.opentracing
-    compile deps.autoservice
-
-    testCompile project(':dd-java-agent:testing')
-    testCompile group: 'io.ratpack', name: 'ratpack-test', version: '1.4.0'
-
-    testCompile group: 'com.squareup.okhttp3', name: 'okhttp', version: '3.6.0'
-  }
 }
 
-if (JavaVersion.current() != JavaVersion.VERSION_1_8) {
-  sourceSets {
-    test {
-      groovy {
-        // These classes use Ratpack which requires Java 8. (Currently also incompatible with Java 9.)
-        exclude '**/RatpackTest.groovy'
-      }
-    }
-  }
+dependencies {
+  main_java8CompileOnly group: 'io.ratpack', name: 'ratpack-core', version: '1.4.0'
+
+  main_java8Compile project(':dd-trace-ot')
+  main_java8Compile project(':dd-java-agent:agent-tooling')
+
+  main_java8Compile deps.bytebuddy
+  main_java8Compile deps.opentracing
+  main_java8Compile deps.autoservice
+
+  compileOnly sourceSets.main_java8.compileClasspath
+
+  compile sourceSets.main_java8.output
+
+  testCompile project(':dd-java-agent:testing')
+  testCompile group: 'io.ratpack', name: 'ratpack-test', version: '1.4.0'
+
+  testCompile group: 'com.squareup.okhttp3', name: 'okhttp', version: '3.6.0'
 }
 
+testJava8Only += '**/RatpackTest.class'
